requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
multidict==6.0.4
numpy==1.24.4
openai==0.27.8
pandas==2.0.3
praw==7.7.1
prawcore==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.8.8
requests==2.31.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.4
websocket-client==1.6.1
yarl==1.9.2
